+++ src/mpol/datasets.py
@@ -1,44 +1,44 @@
 from __future__ import annotations
 
-import copy
 from typing import Any
 
 import numpy as np
 import torch
-import torch.utils.data as torch_ud
 from numpy import floating, integer
 from numpy.typing import ArrayLike, NDArray
 
 from mpol.coordinates import GridCoords
-from mpol.exceptions import WrongDimensionError
 
-from . import utils
-from .constants import *
-from .utils import loglinspace
+from mpol import utils
 
 
 class GriddedDataset(torch.nn.Module):
     r"""
-    Args:
-        coords (GridCoords): an object already instantiated from the GridCoords class. 
-            If providing this, cannot provide ``cell_size`` or ``npix``.
-        vis_gridded (torch complex): the gridded visibility data stored in a "packed" 
-            format (pre-shifted for fft)
-        weight_gridded (torch double): the weights corresponding to the gridded 
-            visibility data, also in a packed format
-        mask (torch boolean): a boolean mask to index the non-zero locations of 
-            ``vis_gridded`` and ``weight_gridded`` in their packed format.
-        nchan (int): the number of channels in the image (default = 1).
-        device (torch.device): the desired device of the dataset. If ``None``, 
-            defaults to current device.
+    Parameters
+    ----------
+    coords : :class:`~mpol.coordinates.GridCoords`
+        If providing this, cannot provide ``cell_size`` or ``npix``.
+    vis_gridded : :class:`torch.Tensor` of :class:`torch.complex128`
+        the gridded visibility data stored in a "packed" format (pre-shifted for fft)
+    weight_gridded : :class:`torch.Tensor` of :class:`torch.double`
+        the weights corresponding to the gridded visibility data,
+        also in a packed format
+    mask : :class:`torch.Tensor` of :class:`torch.bool`
+        a boolean mask to index the non-zero locations of ``vis_gridded`` and
+        ``weight_gridded`` in their packed format.
+    nchan : int
+        the number of channels in the image (default = 1).
 
-    After initialization, the GriddedDataset provides the non-zero cells of the 
-    gridded visibilities and weights as a 1D vector via the following instance 
+
+    After initialization, the GriddedDataset provides the non-zero cells of the
+    gridded visibilities and weights as a 1D vector via the following instance
     variables. This means that any individual channel information has been collapsed.
 
     :ivar vis_indexed: 1D complex tensor of visibility data
+
     :ivar weight_indexed: 1D tensor of weight values
-    
+
+
     If you index the output of the Fourier layer in the same manner using ``self.mask``,
     then the model and data visibilities can be directly compared using a loss function.
     """
@@ -58,9 +58,9 @@ class GriddedDataset(torch.nn.Module):
         self.nchan = nchan
 
         # store variables as buffers of the module
-        self.register_buffer("vis_gridded", torch.tensor(vis_gridded))
-        self.register_buffer("weight_gridded", torch.tensor(weight_gridded))
-        self.register_buffer("mask", torch.tensor(mask))
+        self.register_buffer("vis_gridded", vis_gridded)
+        self.register_buffer("weight_gridded", weight_gridded)
+        self.register_buffer("mask", mask)
         self.vis_gridded: torch.Tensor
         self.weight_gridded: torch.Tensor
         self.mask: torch.Tensor
@@ -73,51 +73,17 @@ class GriddedDataset(torch.nn.Module):
         self.vis_indexed: torch.Tensor
         self.weight_indexed: torch.Tensor
 
-    @classmethod
-    def from_image_properties(
-        cls,
-        cell_size: float,
-        npix: int,
-        *,
-        vis_gridded: torch.Tensor,
-        weight_gridded: torch.Tensor,
-        mask: torch.Tensor,
-        nchan: int = 1,
-        device: torch.device | str | None = None,
-    ):
-        """Alternative method to instantiate a GriddedDataset object from cell_size 
-        and npix.
-
-        Args:
-            cell_size (float): the width of a pixel [arcseconds]
-            npix (int): the number of pixels per image side
-            vis_gridded (torch complex): the gridded visibility data stored in a 
-                "packed" format (pre-shifted for fft)
-            weight_gridded (torch double): the weights corresponding to the gridded 
-                visibility data, also in a packed format
-            mask (torch boolean): a boolean mask to index the non-zero locations of 
-                ``vis_gridded`` and ``weight_gridded`` in their packed format.
-            nchan (int): the number of channels in the image (default = 1).
-        """
-        return cls(
-            coords=GridCoords(cell_size, npix),
-            vis_gridded=vis_gridded,
-            weight_gridded=weight_gridded,
-            mask=mask,
-            nchan=nchan,
-        )
-
     def add_mask(
         self,
         mask: ArrayLike,
     ) -> None:
         r"""
-        Apply an additional mask to the data. Only works as a data limiting operation 
-        (i.e., ``mask`` is more restrictive than the mask already attached 
+        Apply an additional mask to the data. Only works as a data limiting operation
+        (i.e., ``mask`` is more restrictive than the mask already attached
         to the dataset).
 
         Args:
-            mask (2D numpy or PyTorch tensor): boolean mask (in packed format) to 
+            mask (2D numpy or PyTorch tensor): boolean mask (in packed format) to
                 apply to dataset. Assumes input will be broadcast across all channels.
         """
 
@@ -140,15 +106,15 @@ class GriddedDataset(torch.nn.Module):
         self.vis_indexed = self.vis_gridded[self.mask]
         self.weight_indexed = self.weight_gridded[self.mask]
 
-    def forward(self, modelVisibilityCube):
+    def forward(self, modelVisibilityCube: torch.Tensor) -> torch.Tensor:
         """
         Args:
-            modelVisibilityCube (complex torch.tensor): with shape 
-                ``(nchan, npix, npix)`` to be indexed. In "pre-packed" format, as in 
+            modelVisibilityCube (complex torch.tensor): with shape
+                ``(nchan, npix, npix)`` to be indexed. In "pre-packed" format, as in
                 output from :meth:`mpol.fourier.FourierCube.forward()`
 
         Returns:
-            torch complex tensor:  1d torch tensor of indexed model samples collapsed 
+            torch complex tensor:  1d torch tensor of indexed model samples collapsed
                 across cube dimensions.
         """
 
@@ -178,23 +144,24 @@ class GriddedDataset(torch.nn.Module):
         """
         return utils.packed_cube_to_ground_cube(self.mask)
 
+
 class Dartboard:
     r"""
-    A polar coordinate grid relative to a :class:`~mpol.coordinates.GridCoords` object, 
-    reminiscent of a dartboard layout. The main utility of this object is to support 
+    A polar coordinate grid relative to a :class:`~mpol.coordinates.GridCoords` object,
+    reminiscent of a dartboard layout. The main utility of this object is to support
     splitting a dataset along radial and azimuthal bins for k-fold cross validation.
 
     Args:
-        coords (GridCoords): an object already instantiated from the GridCoords class. 
+        coords (GridCoords): an object already instantiated from the GridCoords class.
             If providing this, cannot provide ``cell_size`` or ``npix``.
-        q_edges (1D numpy array): an array of radial bin edges to set the dartboard 
-            cells in :math:`[\mathrm{k}\lambda]`. If ``None``, defaults to 12 
-            log-linearly radial bins stretching from 0 to the :math:`q_\mathrm{max}` 
+        q_edges (1D numpy array): an array of radial bin edges to set the dartboard
+            cells in :math:`[\mathrm{k}\lambda]`. If ``None``, defaults to 12
+            log-linearly radial bins stretching from 0 to the :math:`q_\mathrm{max}`
             represented by ``coords``.
-        phi_edges (1D numpy array): an array of azimuthal bin edges to set the 
-            dartboard cells in [radians], over the domain :math:`[0, \pi]`, which is 
-            also implicitly mapped to the domain :math:`[-\pi, \pi]` to preserve the 
-            Hermitian nature of the visibilities. If ``None``, defaults to 
+        phi_edges (1D numpy array): an array of azimuthal bin edges to set the
+            dartboard cells in [radians], over the domain :math:`[0, \pi]`, which is
+            also implicitly mapped to the domain :math:`[-\pi, \pi]` to preserve the
+            Hermitian nature of the visibilities. If ``None``, defaults to
             8 equal-spaced azimuthal bins stretched from :math:`0` to :math:`\pi`.
     """
 
@@ -217,13 +184,13 @@ class Dartboard:
             # set q edges approximately following inspiration from Petry et al. scheme:
             # https://ui.adsabs.harvard.edu/abs/2020SPIE11449E..1DP/abstract
             # first two bins set to 7m width
-            # after third bin, bin width increases linearly until it is 
+            # after third bin, bin width increases linearly until it is
             # 700m at 16km baseline.
             # From 16m to 16km, bin width goes from 7m to 700m.
             # ---
-            # We aren't doing *quite* the same thing, 
+            # We aren't doing *quite* the same thing,
             # just logspacing with a few linear cells at the start.
-            q_edges = loglinspace(0, self.q_max, N_log=8, M_linear=5)
+            q_edges = utils.loglinspace(0, self.q_max, N_log=8, M_linear=5)
 
         self.q_edges = q_edges
         self.phi_edges = phi_edges
@@ -240,48 +207,21 @@ class Dartboard:
     def q_max(self) -> float:
         return self.coords.q_max
 
-    @classmethod
-    def from_image_properties(
-        cls,
-        cell_size: float,
-        npix: int,
-        q_edges: NDArray[floating[Any]] | None = None,
-        phi_edges: NDArray[floating[Any]] | None = None,
-    ) -> Dartboard:
-        """Alternative method to instantiate a Dartboard object from cell_size
-        and npix.
-
-        Args:
-            cell_size (float): the width of a pixel [arcseconds]
-            npix (int): the number of pixels per image side
-            q_edges (1D numpy array): an array of radial bin edges to set the 
-                dartboard cells in :math:`[\mathrm{k}\lambda]`. If ``None``, defaults 
-                to 12 log-linearly radial bins stretching from 0 to the 
-                :math:`q_\mathrm{max}` represented by ``coords``.
-            phi_edges (1D numpy array): an array of azimuthal bin edges to set the 
-                dartboard cells in [radians], over the domain :math:`[0, \pi]`, which 
-                is also implicitly mapped to the domain :math:`[-\pi, \pi]` to preserve 
-                the Hermitian nature of the visibilities. If ``None``, defaults to 8 
-                equal-spaced azimuthal bins stretched from :math:`0` to :math:`\pi`.
-        """
-        coords = GridCoords(cell_size, npix)
-        return cls(coords, q_edges, phi_edges)
-
     def get_polar_histogram(
         self, qs: NDArray[floating[Any]], phis: NDArray[floating[Any]]
     ) -> NDArray[floating[Any]]:
         r"""
-        Calculate a histogram in polar coordinates, using the bin edges defined by 
+        Calculate a histogram in polar coordinates, using the bin edges defined by
         ``q_edges`` and ``phi_edges`` during initialization.
         Data coordinates should include the points for the Hermitian visibilities.
 
         Args:
             qs: 1d array of q values :math:`[\mathrm{k}\lambda]`
-            phis: 1d array of datapoint azimuth values [radians] (must be the same 
+            phis: 1d array of datapoint azimuth values [radians] (must be the same
                 length as qs)
 
         Returns:
-            2d integer numpy array of cell counts, i.e., how many datapoints fell into 
+            2d integer numpy array of cell counts, i.e., how many datapoints fell into
             each dartboard cell.
         """
 
@@ -297,13 +237,13 @@ class Dartboard:
         self, qs: NDArray[floating[Any]], phis: NDArray[floating[Any]]
     ) -> NDArray[integer[Any]]:
         r"""
-        Return a list of the cell indices that contain data points, using the bin edges 
+        Return a list of the cell indices that contain data points, using the bin edges
         defined by ``q_edges`` and ``phi_edges`` during initialization.
         Data coordinates should include the points for the Hermitian visibilities.
 
         Args:
             qs: 1d array of q values :math:`[\mathrm{k}\lambda]`
-            phis: 1d array of datapoint azimuth values [radians] (must be the same 
+            phis: 1d array of datapoint azimuth values [radians] (must be the same
                 length as qs)
 
         Returns:
@@ -321,8 +261,8 @@ class Dartboard:
         self, cell_index_list: NDArray[integer[Any]]
     ) -> NDArray[np.bool_]:
         r"""
-        Create a boolean mask of size ``(npix, npix)`` (in packed format) corresponding 
-        to the ``vis_gridded`` and ``weight_gridded`` quantities of the 
+        Create a boolean mask of size ``(npix, npix)`` (in packed format) corresponding
+        to the ``vis_gridded`` and ``weight_gridded`` quantities of the
         :class:`~mpol.datasets.GriddedDataset` .
 
         Args: