+++ src/mpol/fourier.py
@@ -15,10 +15,9 @@ from torch import nn
 
 from mpol.exceptions import DimensionMismatchError
 from mpol.images import ImageCube
-from mpol.protocols import MPoLModel
 
-from . import utils
-from .coordinates import GridCoords
+from mpol import utils
+from mpol.coordinates import GridCoords
 
 
 class FourierCube(nn.Module):
@@ -30,64 +29,40 @@ class FourierCube(nn.Module):
     :func:`mpol.losses.nll_gridded`) and a gridded dataset (e.g.,
     :class:`mpol.datasets.GriddedDataset`).
 
-    Args:
-        coords (GridCoords): an object already instantiated from the GridCoords class.
-        persistent_vis (Boolean): should the visibility cube be stored as part of 
-            the module  s `state_dict`? If `True`, the state of the UV grid will be 
-            stored. It is recommended to use `False` for most applications, since the 
-            visibility cube will rarely be a direct parameter of the model.
+    Parameters
+    ----------
+    coords : :class:`~mpol.coordinates.GridCoords`
+        object containing image dimensions
+    persistent_vis : bool
+        should the visibility cube be stored as part of
+        the module  s `state_dict`? If `True`, the state of the UV grid will be
+        stored. It is recommended to use `False` for most applications, since the
+        visibility cube will rarely be a direct parameter of the model.
 
     """
 
     def __init__(self, coords: GridCoords, persistent_vis: bool = False):
         super().__init__()
 
-        # TODO: Is this comment relevant? There was no nchan instantiation
-        # before
-        # ---
-        # we don't want to bother with the nchan argument here, so
-        # we don't use the convenience method _setup_coords
-        # and just do it manually
-
         self.coords = coords
 
         self.register_buffer("vis", None, persistent=persistent_vis)
         self.vis: torch.Tensor
 
-    @classmethod
-    def from_image_properties(
-        cls, cell_size: float, npix: int, persistent_vis: bool = False
-    ) -> FourierCube:
-        """
-        Alternative method for instantiating a FourierCube from ``cell_size`` and
-        ``npix``
-
-        Args:
-            cell_size (float): the width of an image-plane pixel [arcseconds]
-            npix (int): the number of pixels per image side
-            persistent_vis (Boolean): should the visibility cube be stored as part of
-                the modules `state_dict`? If `True`, the state of the UV grid will be
-                stored. It is recommended to use `False` for most applications, since
-                the visibility cube will rarely be a direct parameter of the model.
-
-        Returns:
-            instantiated :class:`mpol.fourier.FourierCube` object.
-        """
-        coords = GridCoords(cell_size, npix)
-        return cls(coords, persistent_vis)
-
     def forward(self, cube: torch.Tensor) -> torch.Tensor:
         """
         Perform the FFT of the image cube on each channel.
 
-        Args:
-            cube (torch.double tensor): a prepacked tensor of shape 
-                ``(nchan, npix, npix)``. For example, an image cube 
-                from ImageCube.forward()
+        Parameters
+        ----------
+        cube : :class:`torch.Tensor` of :class:`torch.double` of shape ``(nchan, npix, npix)``
+            A 'packed' tensor. For example, an image cube from
+            :meth:`mpol.images.ImageCube.forward`
 
-        Returns:
-            torch.complex tensor, of shape ``(nchan, npix, npix)``. The FFT of the
-                image cube, in packed format.
+        Returns
+        -------
+        :class:`torch.Tensor` of :class:`torch.double` of shape ``(nchan, npix, npix)``.
+            The FFT of the image cube, in packed format.
         """
 
         # make sure the cube is 3D
@@ -106,9 +81,11 @@ class FourierCube(nn.Module):
         The visibility cube in ground format cube fftshifted for plotting with
         ``imshow``.
 
-        Returns:
-            (torch.complex tensor, of shape ``(nchan, npix, npix)``): the FFT of the
-                image cube, in sky plane format.
+        Returns
+        -------
+        :class:`torch.Tensor` of :class:`torch.complex128` of shape ``(nchan, npix, npix)``
+            complex-valued FFT of the image cube (i.e., the visibility cube), in 
+            'ground' format.
         """
 
         return utils.packed_cube_to_ground_cube(self.vis)
@@ -119,8 +96,10 @@ class FourierCube(nn.Module):
         The amplitude of the cube, arranged in unpacked format corresponding to the FFT
         of the sky_cube. Array dimensions for plotting given by ``self.coords.vis_ext``.
 
-        Returns:
-            torch.double : 3D amplitude cube of shape ``(nchan, npix, npix)``
+        Returns
+        -------
+        :class:`torch.Tensor` of :class:`torch.double` of shape ``(nchan, npix, npix)``
+            amplitude cube in 'ground' format.
         """
         return torch.abs(self.ground_vis)
 
@@ -130,8 +109,10 @@ class FourierCube(nn.Module):
         The phase of the cube, arranged in unpacked format corresponding to the FFT of
         the sky_cube. Array dimensions for plotting given by ``self.coords.vis_ext``.
 
-        Returns:
-            torch.double : 3D phase cube of shape ``(nchan, npix, npix)``
+        Returns
+        -------
+        :class:`torch.Tensor` of :class:`torch.double` of shape ``(nchan, npix, npix)``
+            phase cube in 'ground' format (:math:`[-\pi,\pi)`).
         """
         return torch.angle(self.ground_vis)
 
@@ -165,22 +146,29 @@ def safe_baseline_constant_meters(
     If this function returns ``True``, then it would be safe to proceed with
     parallelization in the :class:`mpol.fourier.NuFFT` layer via the coil dimension.
 
-    Args:
-        uu (1D np.array): a 1D array of length ``nvis`` array of the u (East-West)
-            spatial frequency coordinate in units of [m]
-        vv (1D np.array): a 1D array of length ``nvis`` array of the v (North-South)
-            spatial frequency coordinate in units of [m]
-        freqs (1D np.array): a 1D array of length ``nchan`` of the channel frequencies,
-            in units of [Hz].
-        coords: a :class:`mpol.coordinates.GridCoords` object which represents the image
-            and uv-grid dimensions.
-        uv_cell_frac (float): the maximum threshold for a change in :math:`u` or
-            :math:`v` spatial frequency across channels, measured as a fraction of the
-            :math:`u,v` cell defined by ``coords``.
+    Parameters
+    ----------
+    uu : 1D np.array 
+        a 1D array of length ``nvis`` array of the u (East-West)
+        spatial frequency coordinate in units of [m]
+    vv : 1D np.array
+        a 1D array of length ``nvis`` array of the v (North-South)
+        spatial frequency coordinate in units of [m]
+    freqs : 1D np.array
+        a 1D array of length ``nchan`` of the channel frequencies,
+        in units of [Hz].
+    coords: :class:`mpol.coordinates.GridCoords` 
+        object which represents the image and uv-grid dimensions.
+    uv_cell_frac : float
+        the maximum threshold for a change in :math:`u` or
+        :math:`v` spatial frequency across channels, measured as a fraction of the
+        :math:`u,v` cell defined by ``coords``.
 
-    Returns:
-        boolean: `True` if it is safe to assume that the baselines are constant with
-            channel (at a tolerance of ``uv_cell_frac``.) Otherwise returns `False`.
+    Returns
+    -------
+    boolean
+        `True` if it is safe to assume that the baselines are constant with
+        channel (at a tolerance of ``uv_cell_frac``.) Otherwise returns `False`.
     """
 
     # broadcast and convert baselines to kilolambda across channel
@@ -299,32 +287,6 @@ class NuFFT(nn.Module):
             im_size=(self.coords.npix, self.coords.npix)
         )
 
-    @classmethod
-    def from_image_properties(
-        cls,
-        cell_size: float,
-        npix: int,
-        nchan: int = 1,
-    ):
-        """
-        Instantiate a :class:`mpol.fourier.NuFFT` object from image properties rather
-        than a :meth:`mpol.coordinates.GridCoords` instance.
-
-        Args:
-            cell_size (float): the width of an image-plane pixel [arcseconds]
-            npix (int): the number of pixels per image side
-            nchan (int): the number of channels in the :class:`mpol.images.ImageCube`.
-                Default = 1.
-
-        Returns:
-            an instance of the :class:`mpol.fourier.NuFFT`
-        """
-        coords = GridCoords(cell_size, npix)
-        return cls(
-            coords,
-            nchan,
-        )
-
     def _klambda_to_radpix(self, klambda: torch.Tensor) -> torch.Tensor:
         """Convert a spatial frequency in units of klambda to 'radians/sky pixel,'
         using the pixel cell_size provided by ``self.coords.dl``.
@@ -382,13 +344,13 @@ class NuFFT(nn.Module):
         vector will influence how TorchKbNufft will perform the operations.
 
         * If ``uu`` and ``vv`` have a 1D shape of (``nvis``), then it will be assumed
-            that the spatial frequencies can be treated as constant with channel. This 
+            that the spatial frequencies can be treated as constant with channel. This
             will result in a ``k_traj`` vector that has shape (``2, nvis``), such that
             parallelization will be across the image cube ``nchan`` dimension using the
             'coil' dimension of the TorchKbNufft package.
         * If the ``uu`` and ``vv`` have a 2D shape of (``nchan, nvis``), then it will
-            be assumed that the spatial frequencies are different for each channel, and 
-            the spatial frequencies provided for each channel will be used. This will 
+            be assumed that the spatial frequencies are different for each channel, and
+            the spatial frequencies provided for each channel will be used. This will
             result in a ``k_traj`` vector that has shape (``nchan, 2, nvis``), such that
             parallelization will be across the image cube ``nchan`` dimension using the
             'batch' dimension of the TorchKbNufft package.
@@ -503,8 +465,8 @@ class NuFFT(nn.Module):
         constant based upon the dimensionality of the ``uu`` and ``vv`` input arguments.
 
         * If ``uu`` and ``vv`` have a shape of (``nvis``), then it will be assumed that
-            the spatial frequencies can be treated as constant with channel (and will 
-            invoke parallelization across the image cube ``nchan`` dimension using the 
+            the spatial frequencies can be treated as constant with channel (and will
+            invoke parallelization across the image cube ``nchan`` dimension using the
             'coil' dimension of the TorchKbNufft package).
         * If the ``uu`` and ``vv`` have a shape of (``nchan, nvis``), then it will be
             assumed that the spatial frequencies are different for each channel, and the
@@ -653,12 +615,12 @@ class NuFFTCached(NuFFT):
 
     * If ``uu`` and ``vv`` have a shape of (``nvis``), then it will be assumed that the
         spatial frequencies can be treated as constant with channel (and will invoke
-        parallelization across the image cube ``nchan`` dimension using the 'coil' 
+        parallelization across the image cube ``nchan`` dimension using the 'coil'
         dimension of the TorchKbNufft package).
     * If the ``uu`` and ``vv`` have a shape of (``nchan, nvis``), then it will be
-        assumed that the spatial frequencies are different for each channel, and the 
-        spatial frequencies provided for each channel will be used (and will invoke 
-        parallelization across the image cube ``nchan`` dimension using the 'batch' 
+        assumed that the spatial frequencies are different for each channel, and the
+        spatial frequencies provided for each channel will be used (and will invoke
+        parallelization across the image cube ``nchan`` dimension using the 'batch'
         dimension of the TorchKbNufft package).
 
     Note that there is no straightforward, computationally efficient way to proceed if
@@ -667,7 +629,7 @@ class NuFFTCached(NuFFT):
     (``nchan, nvis``), such that all channels are padded with bogus :math:`u,v` points
     to have the same length ``nvis``, and you create a boolean mask to keep track of
     which points are valid. Then, when this routine returns data points of shape
-    (``nchan, nvis``), you can use that boolean mask to select only the valid 
+    (``nchan, nvis``), you can use that boolean mask to select only the valid
     :math:`u,v` points.
 
     **Interpolation mode**: You may choose the type of interpolation mode that KbNufft
@@ -736,19 +698,6 @@ class NuFFTCached(NuFFT):
             self.register_buffer("imag_interp_mat", imag_interp_mat)
             self.real_interp_mat: torch.Tensor
             self.imag_interp_mat: torch.Tensor
-
-    @classmethod
-    def from_image_properties(
-        cls,
-        cell_size,
-        npix,
-        uu,
-        vv,
-        nchan=1,
-        sparse_matrices=True,
-    ):
-        coords = GridCoords(cell_size, npix)
-        return cls(coords, uu, vv, nchan, sparse_matrices)
 
     def forward(self, cube):
         r"""
@@ -903,7 +852,7 @@ def get_vis_residuals(model, u_true, v_true, V_true, return_Vmod=False, channel=
     """
     nufft = NuFFT(coords=model.coords, nchan=model.nchan)
 
-    vis_model = nufft(model.icube().to("cpu"), u_true, v_true)  # TODO: remove 'to' call
+    vis_model = nufft(model.icube.cube.to("cpu"), u_true, v_true)  # TODO: remove 'to' call
     # convert to numpy, select channel
     vis_model = vis_model.detach().numpy()[channel]